refactor(command): provider structure directory
