+++ stubs/application/provider.stub
@@ -8,8 +8,6 @@ class {{ class }} extends ServiceProvider
 {
     /**
      * Register services.
-     *
-     * @return void
      */
     public function register(): void
     {
@@ -18,8 +16,6 @@ class {{ class }} extends ServiceProvider
 
     /**
      * Bootstrap services.
-     *
-     * @return void
      */
     public function boot(): void
     {
